Add test for "major tags only", expand conditionals
The construct

  [[ condition ]] && something

resulted in the tests failing because of non-zero exit status. Since the
tagtable function updates a nameref, it can't be run within a Bats "run"
command, and because it doesn't, a non-zero exit status is interpreted
as a failed assertion. This commit expands the constructs to standard
if/then.